unified packaging: initial implementation debian/control.in
The aim is to share the packaging between wine (stable) and
wine-development, based upon the already existing filenames with VERSION.

If there are necessary differences, they are handled within the files
(e.g. control.in) by adding the tags #stable# and #development# in the
differing lines.

In this first step I kept most of the differences between stable and
development. Already incorporated changes:
- wine32|64-dev-tools are renamed to wine32|64-tools (without -dev)
- the preloader packages are new for wine stable