Revert "Move files in /usr/include to arch specific directory."
This reverts commit c5d3c08e3069435c56606d5eeb2bb9308f449239.

This endianess issue should already be fixed upstream in Wine 1.7.16 by:

commit b8ad360bbdb69593beace414199d5e37bf83f843
Author: Alexandre Julliard <julliard@winehq.org>
Date: Thu Apr 3 11:31:59 2014 +0200
    make_xftmpl: Always output data in little-endian format.

Disclaimer: Actual programming is not my field of expertise, thus I can't
verify if this commit really fixes this. Also I'm not aware of any place
to check the current MD5 sums on powerpc.

Thanks Michael Gilbert for the hint.