Don't set a default pulseaudio latency.
This was originally added to avoid stuttering (#581624). With the new
pulse audio driver it isn't needed anymore. Instead it may cause
applications to not play sound because the pulse buffer may be smaller
then requested by Wine.
Confirmed and tested, both that applications don't start to stutter
again, and that sound works again, see
https://bugs.winehq.org/show_bug.cgi?id=39744

Closes: #807403