Move Wine binaries to common directory.
This fixes the WoW64 setup for wine-development (closes: #813475).

Wine first looks in its bindir for the loaders and the server.
This bindir seems to be the dirname of the calling executable. So
if wine, wine64 and wineserver are in the same directory, they
are automatically found.

This bindir needs to be in the same file hierarchy level as the
bindir specified on package build, to prevent broken wine
internal links like /usr/lib/wine-development/../../../share/
wine-development/wine/wine.inf.

Therefore:
Move everything in /usr/lib/wineVERSION to new subdir wine.
Move adjusted wine and wineserver scripts to this dir, and create
links in /usr/bin.
Don't specify WINELOADER and WINESERVER explicitly.