Install the 32-bit wine loader directly as bindir/wine.
Otherwise the preloader, if installed, seems to have problems:
  /usr/lib/wine-development/wine: failed to read ELF header

Restore the wine script as /usr/bin/wineVERSION. If the 32-bit
wine is missing, export the WINELOADER wine64 to avoid a fallback
to /usr/bin/wine as loader.

Drop the wine32 link in /usr/bin, since we don't have a binary
named wine32 anymore.