Add generated lists of shared libs to libwine's dependencies.
Closes: #823991 (stable)

They are used in the Depends (${dlopen:Depends},) and Recommends
(${dlopen:Recommends}) of libwineVERSION.

Remove hardcoded duplicates. Next to libfontconfig1 and
libxcursor1 remove all dependencies recently added in commit
3afa464 (Add missing runtime dependencies to libwine recommends).
Not all of them are generated now.

Compared to 1.9.8-1 the following dependencies are added (some
of them with a minimum version requirement):
 libcups2
 libdbus-1-3
 libgl1-mesa-glx|libgl1
 libgnutls30
 libgsm1
 libjpeg62-turbo
 libodbc1
 libosmesa6
 libpng16-16
 libtiff5
 libxcomposite1
 libxi6
 libxinerama1
 libxrandr2
 libxrender1
 libxslt1.1
 libxxf86vm1