Drop useless workaround for one-arch-only binNMUs.
binNMUs need to be in sync across architectures anyway, because
libwineVERSION automatically breaks every other binary:Version of
itself. Note that this isn't in DEBIAN/control, but may be
observed e.g. in aptitude.

Drop this workaround to be in line with the next changes which
would be really ugly if they replicated this.